Add new `shopapp` module to the Django project.
This commit creates a new module called `shopapp` and integrates it into the primary Django application. This module contains several new files, including models, views, admin, tests, and urls. In addition, the `ShopappConfig` was added into the Django settings file and the routing for `shopapp` is set up in `urls.py` file.